beverly_hills/requirements.txt
scipy==1.4.1
xgboost==1.2.0
numpy==1.18.5
# tensorflow==2.2.0
scikit_learn==0.23.2
websocket-client==0.57.0
autobahn==20.12.3